finish promotion service

+++ GameShop-Backend/src/main/java/ca/mcgill/ecse321/GameShop/controller/PromotionController.java
@@ -55,7 +55,7 @@ public class PromotionController {
      * @return The updated promotion information.
      */
     @PutMapping("/{promotionId}")
-    public PromotionResponseDto updatePromotion(@PathVariable Long promotionId,
+    public PromotionResponseDto updatePromotion(@PathVariable Integer promotionId,
             @RequestBody PromotionRequestDto promotionRequestDto) {
         Promotion promotion = promotionService.updatePromotion(promotionId, promotionRequestDto);
         return new PromotionResponseDto(promotion);
@@ -67,7 +67,7 @@ public class PromotionController {
      * @param promotionId The ID of the promotion to delete.
      */
     @DeleteMapping("/{promotionId}")
-    public void deletePromotion(@PathVariable Long promotionId) {
+    public void deletePromotion(@PathVariable Integer promotionId) {
         promotionService.deletePromotion(promotionId);
     }
 
@@ -78,7 +78,7 @@ public class PromotionController {
      * @return A list of promotions associated with the given game.
      */
     @GetMapping("/game/{gameId}")
-    public PromotionListDto getPromotionsByGame(@PathVariable Long gameId) {
+    public PromotionListDto getPromotionsByGame(@PathVariable Integer gameId) {
         List<PromotionResponseDto> promotionDtos = new ArrayList<PromotionResponseDto>();
         for (Promotion promotion : promotionService.getPromotionsByGameId(gameId)) {
             promotionDtos.add(new PromotionResponseDto(promotion));
@@ -93,7 +93,7 @@ public class PromotionController {
      * @return A list of promotions associated with the given category.
      */
     @GetMapping("/category/{categoryId}")
-    public PromotionListDto getPromotionsByCategory(@PathVariable Long categoryId) {
+    public PromotionListDto getPromotionsByCategory(@PathVariable Integer categoryId) {
         List<PromotionResponseDto> promotionDtos = new ArrayList<PromotionResponseDto>();
         for (Promotion promotion : promotionService.getPromotionsByCategoryId(categoryId)) {
             promotionDtos.add(new PromotionResponseDto(promotion));
